0x02 tasks

+++ 0x02-functions_nested_loops/main.h
@@ -11,10 +11,11 @@ int _isalpha(int c);
 int _islower(int c);
 int print_sign(int);
 int _abs(int);
-int print_last_digit(int);
+int print_last_digit(long int);
 void jack_bauer(void);
 void times_table(void);
-int add(int, int);void print_to_98(int n);
+int add(int, int);
+void print_to_98(int n);
 void print_times_table(int n);
 
 #endif
